requirements.txt
streamlit
langchain-ollama
langchain-openai
langchain-community
langchain-core
langgraph
fpdf
datetime
python-dotenv
